as I said I can hardly do more than 1-2 hours to work on it. I also think to aim at too many changes, which would be justified by the need, but hapert in the implementation.
What I am currently trying to do is to isolate the handling for downloading into individual functions, things like:

Enable aria2 multisocket automatically if:

    aria2 is installed: checkDownloaders () in doOptions.R
    dlmethod = aria2, multisocket = TRUE defined in the new ~ / .MODISopts.R
    genString () $ remotePath has more than one path.

If multisocket is used then both remothPath's become a paste()'d into one string.

Then we should also isolate the credentials handing (having a handler who acts properly depending on the condition). In general, the .netrc file is very pleasant, since you do not have to bother about auth in the functions themselves. But it would be silly if .netrc is only partially supported. As an alternative, a string (or command) making function that creates situationdepending the correct string/command. Generally, for this I would not be too much 'revolutionizing' because of the time pressure.